Decompile DRA func_8012F894
